+++ package.json
@@ -22,6 +22,7 @@
     "react-js-pagination": "^3.0.3",
     "react-paginate": "^8.2.0",
     "react-scripts": "5.0.1",
+    "react-spinners": "^0.13.8",
     "scss": "^0.2.4",
     "styled-components": "^6.0.1",
     "use-debounce": "^9.0.4",
