--- server/selectors.ts
@@ -1,13 +1,0 @@
-import { ISelectors } from "./types";
-
-export const ebaySelectors: ISelectors = {
-  price: ".x-price-primary > span > .ux-textspans",
-  title: ".x-item-title > h1 > span",
-  img: ".image img",
-};
-
-export const amazonSelectors: ISelectors = {
-  title: "h1 > #productTitle",
-  price: ".a-price > span",
-  img: ".imgTagWrapper img",
-};
