+++ src/Common/Button.tsx
@@ -1,0 +1,10 @@
+import React, { ButtonHTMLAttributes } from "react";
+import { StyledButton } from "./Button.styled";
+
+type ButtonProps = ButtonHTMLAttributes<HTMLButtonElement>;
+
+const Button: React.FC<ButtonProps> = ({ children, ...rest }) => {
+  return <StyledButton {...rest}>{children}</StyledButton>;
+};
+
+export default Button;
