+++ src/ScrapingApp/Items/Items.tsx
@@ -1,4 +1,4 @@
-import React, { useEffect, useState } from "react";
+import React from "react";
 import styled from "styled-components";
 import ItemRow, { ItemRowType } from "../ItemRow/ItemRow";
